UIDEXP-113 create hook for edit and duplicate
